fix: add explicit UInt8() casts in _parse_int_label() (closes #283) (#284)
Mojo deprecated implicit Int→UInt8 narrowing conversions. The two
comparisons `bytes[0] == ord('-')` and `bytes[0] == ord('+')` triggered
the warning; wrap the ord() results in UInt8() to make the conversions
explicit before implicit narrowing is removed from the compiler.

+++ bison/indexing.mojo
@@ -31,10 +31,10 @@ def _parse_int_label(label: String) raises -> Int:
     var bytes = label.as_bytes()
     var start = 0
     var negative = False
-    if bytes[0] == ord('-'):
+    if bytes[0] == UInt8(ord('-')):
         negative = True
         start = 1
-    elif bytes[0] == ord('+'):
+    elif bytes[0] == UInt8(ord('+')):
         start = 1
     if start >= n:
         raise Error("loc: invalid row label: " + label)
